- add globalThreadsTab to SYSTEM_IDENTIFIERS - save the lastview globalThreadsTab to the database after unmounting the threads view - default to all on first opening of the threads view
